requirements.txt
numpy==1.20.1
Pillow==8.1.1
requests==2.25.1
pymongo==3.11.3
pytest==6.2.2
fastapi==0.95.1
uvicorn==0.21.1
mysql-connector-python==8.0.33
python-multipart==0.0.6